Rename several multiple-symbol files to snake-case
