+++ src/content-scripts/class-consts.ts
