+++ src/content-scripts/components/viewer/viewer-elements/UserActionDialog.tsx
@@ -4,7 +4,7 @@ import { connect } from 'react-redux';
 import styles from '../report-editor/ReportEditor.scss';
 import { changeViewerReportEditorOpen } from 'content-scripts/store/widgets/actions';
 import { AnnotationAPIModel } from 'content-scripts/api/annotations';
-import { PPScopeClass } from 'content-scripts/class_consts';
+import { PPScopeClass } from 'content-scripts/class-consts';
 import ReportEditor from '../report-editor/ReportEditor';
 import ppGA from 'common/pp-ga';
