+++ src/content-scripts/utils/TextSelector.ts
@@ -1,7 +1,7 @@
 import { Range as XPathRange } from 'xpath-range';
 import $ from 'jquery';
 import _isEqual from 'lodash/isEqual';
-import { PPHighlightClass } from 'content-scripts/class_consts';
+import { PPHighlightClass } from 'content-scripts/class-consts';
 // More on xpath-range here: https://github.com/opengovfoundation/xpath-range
 // Wondering what's inside? See https://github.com/opengovfoundation/xpath-range/blob/master/src/range.coffee#L227
 import { annotationRootNode } from '../main';
