+++ src/content-scripts/utils/index.ts
@@ -1,3 +1,4 @@
 export { default as Highlighter } from './Highlighter';
 export { default as TextSelector } from './TextSelector';
 export { default as DragTracker } from './DragTracker';
+export { default as mousePosition } from './mousePosition';
